+++ unique_toolkit/tests/language_model/test_language_models_service_unit.py
@@ -3,7 +3,7 @@ from unittest.mock import patch
 import pytest
 import unique_sdk
 
-from unique_toolkit.chat.state import ChatState
+from tests.test_obj_factory import get_event_obj
 from unique_toolkit.content.schemas import ContentChunk
 from unique_toolkit.language_model.infos import LanguageModelName
 from unique_toolkit.language_model.schemas import (
@@ -42,13 +42,13 @@ mock_tool = LanguageModelTool(
 class TestLanguageModelServiceUnit:
     @pytest.fixture(autouse=True)
     def setup(self):
-        self.chat_state = ChatState(
+        self.event = get_event_obj(
             user_id="test_user",
             company_id="test_company",
             assistant_id="test_assistant",
             chat_id="test_chat",
         )
-        self.service = LanguageModelService(self.chat_state)
+        self.service = LanguageModelService(self.event)
 
     def test_complete(self):
         with patch.object(unique_sdk.ChatCompletion, "create") as mock_create: