+++ unique_toolkit/unique_toolkit/chat/__init__.py
@@ -1,7 +1,6 @@
 from .schemas import ChatMessage as ChatMessage
 from .schemas import ChatMessageRole as ChatMessageRole
 from .service import ChatService as ChatService
-from .state import ChatState as ChatState
 from .utils import (
     convert_chat_history_to_injectable_string as convert_chat_history_to_injectable_string,
 )