+++ unique_toolkit/unique_toolkit/content/service.py
@@ -8,7 +8,7 @@ import requests
 import unique_sdk
 
 from unique_toolkit._common._base_service import BaseService
-from unique_toolkit.chat.state import ChatState
+from unique_toolkit.app.schemas import Event
 from unique_toolkit.content.schemas import (
     Content,
     ContentChunk,
@@ -23,12 +23,12 @@ class ContentService(BaseService):
     Provides methods for searching, downloading and uploading content in the knowledge base.
 
     Attributes:
-        state (ChatState): The chat state.
+        event (Event): The Event object.
         logger (Optional[logging.Logger]): The logger. Defaults to None.
     """
 
-    def __init__(self, state: ChatState, logger: Optional[logging.Logger] = None):
-        super().__init__(state, logger)
+    def __init__(self, event: Event, logger: Optional[logging.Logger] = None):
+        super().__init__(event, logger)
 
     DEFAULT_SEARCH_LANGUAGE = "english"
 
@@ -62,9 +62,9 @@ class ContentService(BaseService):
 
         try:
             searches = unique_sdk.Search.create(
-                user_id=self.state.user_id,
-                company_id=self.state.company_id,
-                chatId=self.state.chat_id,
+                user_id=self.event.user_id,
+                company_id=self.event.company_id,
+                chatId=self.event.payload.chat_id,
                 searchString=search_string,
                 searchType=search_type.name,
                 scopeIds=scope_ids,
@@ -116,9 +116,9 @@ class ContentService(BaseService):
 
         try:
             searches = await unique_sdk.Search.create_async(
-                user_id=self.state.user_id,
-                company_id=self.state.company_id,
-                chatId=self.state.chat_id,
+                user_id=self.event.user_id,
+                company_id=self.event.company_id,
+                chatId=self.event.payload.chat_id,
                 searchString=search_string,
                 searchType=search_type.name,
                 scopeIds=scope_ids,
@@ -156,9 +156,9 @@ class ContentService(BaseService):
         """
         try:
             contents = unique_sdk.Content.search(
-                user_id=self.state.user_id,
-                company_id=self.state.company_id,
-                chatId=self.state.chat_id,
+                user_id=self.event.user_id,
+                company_id=self.event.company_id,
+                chatId=self.event.payload.chat_id,
                 # TODO add type parameter
                 where=where,  # type: ignore
             )
@@ -183,9 +183,9 @@ class ContentService(BaseService):
         """
         try:
             contents = await unique_sdk.Content.search_async(
-                user_id=self.state.user_id,
-                company_id=self.state.company_id,
-                chatId=self.state.chat_id,
+                user_id=self.event.user_id,
+                company_id=self.event.company_id,
+                chatId=self.event.payload.chat_id,
                 # TODO add type parameter
                 where=where,  # type: ignore
             )
@@ -325,8 +325,8 @@ class ContentService(BaseService):
                     "mimeType": input.mime_type,
                 }
             content = unique_sdk.Content.upsert(
-                user_id=self.state.user_id,
-                company_id=self.state.company_id,
+                user_id=self.event.user_id,
+                company_id=self.event.company_id,
                 input=input_json,  # type: ignore
                 fileUrl=content_url,
                 scopeId=scope_id,
@@ -374,8 +374,8 @@ class ContentService(BaseService):
         headers = {
             "x-api-version": unique_sdk.api_version,
             "x-app-id": unique_sdk.app_id,
-            "x-user-id": self.state.user_id,
-            "x-company-id": self.state.company_id,
+            "x-user-id": self.event.user_id,
+            "x-company-id": self.event.company_id,
             "Authorization": "Bearer %s" % (unique_sdk.api_key,),
         }